+++ utoipa/src/openapi/path.rs
@@ -15,9 +15,13 @@ use super::{
 };
 
 #[cfg(not(feature = "preserve_path_order"))]
-pub(super) type PathsMap<K, V> = std::collections::BTreeMap<K, V>;
+#[allow(missing_docs)]
+#[doc(hidden)]
+pub type PathsMap<K, V> = std::collections::BTreeMap<K, V>;
 #[cfg(feature = "preserve_path_order")]
-pub(super) type PathsMap<K, V> = indexmap::IndexMap<K, V>;
+#[allow(missing_docs)]
+#[doc(hidden)]
+pub type PathsMap<K, V> = indexmap::IndexMap<K, V>;
 
 builder! {
     PathsBuilder;